requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
xgboost==2.0.0
joblib==1.3.2
python-multipart==0.0.6
pyjwt==2.8.0
slowapi==0.1.9
pydantic==2.5.0